Move ChatList component to dedicated file
